# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Multi-Agent Chat System API
  description: |
    A powerful multi-agent AI system for collaborative software development, featuring 
    specialized agents for architecture, coding, testing, security, and DevOps, with 
    advanced NLU capabilities and intelligent task orchestration.
  version: 2.0.0
  contact:
    name: API Support
    url: https://github.com/anthropics/claude-code/issues
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5006/api
    description: Development server
  - url: https://api.example.com/api/v2
    description: Production server (v2)
  - url: https://api.example.com/api/v1
    description: Production server (v1 - deprecated)
security:
  - ApiKeyAuth: []
...
paths:
  /agents/list:
    get:
      tags:
        - Agents
      summary: List available agents
      description: Returns a list of all available AI agents with their capabilities
      operationId: listAgents
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /agents/chat/{agent_id}:
    post:
      tags:
        - Agents
      summary: Chat with specific agent
      description: Send a message to a specific AI agent and receive a response
      operationId: chatWithAgent
      parameters:
        - name: agent_id
          in: path
          required: true
          description: Unique identifier of the agent
          schema:
            type: string
            example: developer_01
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /agents/collaborate:
    post:
      tags:
        - Agents
      summary: Multi-agent collaboration
      description: Execute a task using multiple agents working together
      operationId: collaborateAgents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollaborateRequest'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollaborateResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Message to send to the agent
          example: "Create a REST API endpoint for user authentication"
        enhance_prompt:
          type: boolean
          default: true
          description: Whether to enhance the prompt for better results
        include_context:
          type: boolean
          default: true
          description: Include conversation context
        session_id:
          type: string
          description: Optional session ID for conversation continuity
    ChatResponse:
      type: object
      properties:
        success:
          type: boolean
        response:
          type: string
          description: Agent's response
        agent:
          type: string
          description: Agent ID that responded
        timestamp:
          type: string
          format: date-time
        enhanced:
          type: boolean
          description: Whether prompt was enhanced
        original_message:
          type: string
          description: Original message if enhanced
...
    CollaborateRequest:
      type: object
      required:
        - task
      properties:
        task:
          type: string
        agents:
          type: array
          items:
            type: string
        workflow:
          type: string
          enum: [sequential, parallel]
          default: sequential
        context:
          type: object
          additionalProperties: true
    CollaborateResponse:
      type: object
      properties:
        success:
          type: boolean
        task_id:
          type: string
        status:
          type: string
        agents_assigned:
          type: array
          items:
            type: string
        estimated_completion:
          type: string
          format: date-time
...
    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              role:
                type: string
              description:
                type: string
              capabilities:
                type: array
                items:
                  type: string
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "INVALID_REQUEST"
            message:
              type: string
              example: "Task description is required"
            details:
              type: object
              additionalProperties: true
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...